requirements.txt
pandas>=2.0.0
duckdb>=0.9.0
python-dotenv==1.0.1
boto3==1.34.34
PyMuPDF>=1.22.0
Pillow>=10.0.0
flask==3.0.2
pytest==7.4.0
pytest-mock==3.14.0
pytest-cov==4.1.0
requests==2.31.0 